+++ src/scoring.js
@@ -4,11 +4,14 @@ import { Box, Point } from './foundation.js';
 import * as geom from './geom.js';
 
 export function closingPenalty(cd, opt) {
-    return (cd > (opt.scoring.closingDistanceFree || 0) ? cd : 0);
+    return opt.scoring.rounding(
+        (cd > (opt.scoring.closingDistanceFree || 0) ? cd : 0));
 }
 
 export function closingWithLimit(distance, opt) {
-    return Math.max(opt.scoring.closingDistanceFixed || 0, distance * (opt.scoring.closingDistanceRelative || 0));
+    return opt.scoring.rounding(
+        Math.max(opt.scoring.closingDistanceFixed || 0,
+            distance * (opt.scoring.closingDistanceRelative || 0)));
 }
 
 /*eslint no-unused-vars: ["error", { "args": "none" }]*/
@@ -21,10 +24,11 @@ export function closingWithPenalty(distance, opt) {
 export function boundDistance3Points(ranges, boxes, opt) {
     const pin = geom.findFurthestPointInSegment(opt.launch, ranges[0].start, boxes[0], opt);
     const pout = geom.findFurthestPointInSegment(ranges[2].end, opt.landing, boxes[2], opt);
-    const maxDistance = geom.maxDistanceNRectangles([pin, boxes[0], boxes[1], boxes[2], pout]);
+    const maxDistance = opt.scoring.rounding(
+        geom.maxDistanceNRectangles([pin, boxes[0], boxes[1], boxes[2], pout]));
     if (maxDistance < (opt.scoring.minDistance || 0))
         return 0;
-    return maxDistance * opt.scoring.multiplier;
+    return opt.scoring.rounding(maxDistance * opt.scoring.multiplier);
 }
 
 // Score of a 3TP distance flight with all 3 points selected
@@ -33,10 +37,22 @@ export function scoreDistance3Points(tp, opt) {
     const pin = geom.findFurthestPointInSegment(opt.launch, tp[0].r, tp[0], opt);
     const pout = geom.findFurthestPointInSegment(tp[2].r, opt.landing, tp[2], opt);
     const all = [pin, tp[0], tp[1], tp[2], pout];
-    for (let i = 0; i < all.length - 1; i++)
-        distance += all[i].distanceEarth(all[i + 1]);
-    const score = distance >= (opt.scoring.minDistance || 0) ? distance * opt.scoring.multiplier : 0;
-    return { distance, score, tp: tp, ep: { start: pin, finish: pout } };
+    const legs = [
+        {name: 'START : TP1'},
+        {name: 'TP1 : TP2'},
+        {name: 'TP2 : TP3'},
+        {name: 'TP3 : FINISH'}
+    ];
+    for (let i = 0; i < all.length - 1; i++) {
+        legs[i].d = opt.scoring.rounding(all[i].distanceEarth(all[i + 1]));
+        distance += legs[i].d;
+        legs[i].start = all[i];
+        legs[i].finish = all[i+1];
+    }
+    const score = distance >= (opt.scoring.minDistance || 0) ?
+        opt.scoring.rounding(distance * opt.scoring.multiplier) :
+        0;
+    return { distance, score, tp: tp, ep: { start: pin, finish: pout }, legs };
 }
 
 // Upper limit for a FAI triangle with vertices somewhere in boxes,
@@ -44,19 +60,21 @@ export function scoreDistance3Points(tp, opt) {
 function maxFAIDistance(maxTriDistance, boxes, opt) {
     // smallest triangle with we could possibly have for these boxes
     const minTriDistance = geom.minDistance3Rectangles(boxes, (i, j, k) => {
-        return i.distanceEarth(j) + j.distanceEarth(k) + k.distanceEarth(i);
+        return opt.scoring.rounding(i.distanceEarth(j)) +
+            opt.scoring.rounding(j.distanceEarth(k)) +
+            opt.scoring.rounding(k.distanceEarth(i));
     });
 
     if (maxTriDistance < minTriDistance)
         return 0;
 
     // biggest possible leg for each side
-    const maxAB = geom.maxDistance2Rectangles([boxes[0], boxes[1]]);
-    const maxBC = geom.maxDistance2Rectangles([boxes[1], boxes[2]]);
-    const maxCA = geom.maxDistance2Rectangles([boxes[2], boxes[0]]);
+    const maxAB = opt.scoring.rounding(geom.maxDistance2Rectangles([boxes[0], boxes[1]]));
+    const maxBC = opt.scoring.rounding(geom.maxDistance2Rectangles([boxes[1], boxes[2]]));
+    const maxCA = opt.scoring.rounding(geom.maxDistance2Rectangles([boxes[2], boxes[0]]));
 
     // our FAI triangle is limited to maxDistance
-    const maxDistance = Math.min(maxAB, maxBC, maxCA) / opt.scoring.minSide;
+    const maxDistance = opt.scoring.rounding(Math.min(maxAB, maxBC, maxCA) / opt.scoring.minSide);
 
     // Is the maximum FAI triangle smaller than the minimum possible triangle?
     if (maxDistance < minTriDistance)
@@ -69,12 +87,12 @@ function maxFAIDistance(maxTriDistance, boxes, opt) {
 // maxTriDistance is the upper limit of the unconstrained flat triangle
 function maxTRIDistance(maxTriDistance, boxes, opt) {
     // smallest possible leg for each side
-    const minAB = geom.minDistance2Rectangles([boxes[0], boxes[1]]);
-    const minBC = geom.minDistance2Rectangles([boxes[1], boxes[2]]);
-    const minCA = geom.minDistance2Rectangles([boxes[2], boxes[0]]);
+    const minAB = opt.scoring.rounding(geom.minDistance2Rectangles([boxes[0], boxes[1]]));
+    const minBC = opt.scoring.rounding(geom.minDistance2Rectangles([boxes[1], boxes[2]]));
+    const minCA = opt.scoring.rounding(geom.minDistance2Rectangles([boxes[2], boxes[0]]));
 
     // our constrained triangle cannot be smaller then minDistance
-    const minDistance = Math.max(minAB, minBC, minCA) / opt.scoring.maxSide;
+    const minDistance = opt.scoring.rounding(Math.max(minAB, minBC, minCA) / opt.scoring.maxSide);
 
     // Is the minimum constrained triangle bigger than the maximum possible triangle?
     if (minDistance > maxTriDistance)
@@ -89,9 +107,11 @@ function maxTRIDistance(maxTriDistance, boxes, opt) {
 export function boundOpenTriangle(ranges, boxes, opt) {
     const pin = geom.findFurthestPointInSegment(opt.launch, ranges[0].start, boxes[0], opt);
     const pout = geom.findFurthestPointInSegment(ranges[2].end, opt.landing, boxes[2], opt);
-    const maxD3PDistance = geom.maxDistanceNRectangles([pin, boxes[0], boxes[1], boxes[2], pout]);
+    const maxD3PDistance = opt.scoring.rounding(geom.maxDistanceNRectangles([pin, boxes[0], boxes[1], boxes[2], pout]));
     const maxTriDistance = geom.maxDistance3Rectangles(boxes, (i, j, k) => {
-        return i.distanceEarth(j) + j.distanceEarth(k) + k.distanceEarth(i);
+        return opt.scoring.rounding(i.distanceEarth(j)) +
+            opt.scoring.rounding(j.distanceEarth(k)) +
+            opt.scoring.rounding(k.distanceEarth(i));
     });
     if (maxTriDistance < (opt.scoring.minDistance || 0))
         return 0;
@@ -109,16 +129,16 @@ export function boundOpenTriangle(ranges, boxes, opt) {
         cp = geom.isTriangleClosed(ranges[0].end, ranges[2].start, maxTriDistance, opt);
         if (!cp)
             return 0;
-        return (maxD3PDistance - closingPenalty(cp.d, opt)) * opt.scoring.multiplier;
+        return opt.scoring.rounding((maxD3PDistance - closingPenalty(cp.d, opt)) * opt.scoring.multiplier);
     }
 
-    return maxD3PDistance * opt.scoring.multiplier;
+    return opt.scoring.rounding(maxD3PDistance * opt.scoring.multiplier);
 }
 
 export function scoreOpenTriangle(tp, opt) {
-    const d0 = tp[0].distanceEarth(tp[1]);
-    const d1 = tp[1].distanceEarth(tp[2]);
-    const d2 = tp[2].distanceEarth(tp[0]);
+    const d0 = opt.scoring.rounding(tp[0].distanceEarth(tp[1]));
+    const d1 = opt.scoring.rounding(tp[1].distanceEarth(tp[2]));
+    const d2 = opt.scoring.rounding(tp[2].distanceEarth(tp[0]));
     const triDistance = d0 + d1 + d2;
 
     if (triDistance < (opt.scoring.minDistance || 0))
@@ -134,23 +154,34 @@ export function scoreOpenTriangle(tp, opt) {
     if (!cp)
         return { score: 0 };
 
-    let d3pDistance = 0;
     const pin = geom.findFurthestPointInSegment(opt.launch, tp[0].r, tp[0], opt);
     const pout = geom.findFurthestPointInSegment(tp[2].r, opt.landing, tp[2], opt);
     const all = [pin, tp[0], tp[1], tp[2], pout];
-    for (let i = 0; i < all.length - 1; i++)
-        d3pDistance += all[i].distanceEarth(all[i + 1]);
+    const legs = [
+        {name: 'START : TP1'},
+        {name: 'TP1 : TP2'},
+        {name: 'TP2 : TP3'},
+        {name: 'TP3 : FINISH'}
+    ];
+    let distance = 0;
+    for (let i = 0; i < all.length - 1; i++) {
+        legs[i].d = opt.scoring.rounding(all[i].distanceEarth(all[i + 1]));
+        distance += legs[i].d;
+        legs[i].start = all[i];
+        legs[i].finish = all[i+1];
+    }
 
-    const distance = d3pDistance;
     const score = distance * opt.scoring.multiplier - closingPenalty(cp.d, opt);
-    return { distance, score, tp: tp, ep: { start: pin, finish: pout }, cp };
+    return { distance, score, tp: tp, ep: { start: pin, finish: pout }, cp, legs };
 }
 /* c8 ignore stop */
 
 // Upper limit for a flat triangle with vertices somewhere in boxes
 export function boundTriangle(ranges, boxes, opt) {
     const maxTriDistance = geom.maxDistance3Rectangles(boxes, (i, j, k) => {
-        return i.distanceEarth(j) + j.distanceEarth(k) + k.distanceEarth(i);
+        return opt.scoring.rounding(i.distanceEarth(j)) +
+            opt.scoring.rounding(j.distanceEarth(k)) +
+            opt.scoring.rounding(k.distanceEarth(i));
     });
 
     if (maxTriDistance < (opt.scoring.minDistance || 0))
@@ -170,32 +201,40 @@ export function boundTriangle(ranges, boxes, opt) {
         cp = geom.isTriangleClosed(ranges[0].end, ranges[2].start, maxDistance, opt);
         if (!cp)
             return 0;
-        return (maxDistance - closingPenalty(cp.d, opt)) * opt.scoring.multiplier;
+        return opt.scoring.rounding((maxDistance - closingPenalty(cp.d, opt)) * opt.scoring.multiplier);
     }
 
     // Ranges overlap - bounding is impossible at this stage
-    return maxDistance * opt.scoring.multiplier;
+    return opt.scoring.rounding(maxDistance * opt.scoring.multiplier);
 }
 
 // Score a triangle once all 3 points have been selected
 export function scoreTriangle(tp, opt) {
-    const d0 = tp[0].distanceEarth(tp[1]);
-    const d1 = tp[1].distanceEarth(tp[2]);
-    const d2 = tp[2].distanceEarth(tp[0]);
-    const distance = d0 + d1 + d2;
+    let distance = 0;
+    const legs = [
+        {name: 'TP1 : TP2'},
+        {name: 'TP2 : TP3'},
+        {name: 'TP3 : TP1'}
+    ];
+    for (let i = 0; i < tp.length; i++) {
+        legs[i].d = opt.scoring.rounding(tp[i].distanceEarth(tp[(i + 1) % tp.length]));
+        distance += legs[i].d;
+        legs[i].start = tp[i];
+        legs[i].finish = tp[(i + 1) % tp.length];
+    }
 
     if (distance < (opt.scoring.minDistance || 0))
         return { score: 0 };
 
     if (opt.scoring.minSide !== undefined) {
         const minSide = opt.scoring.minSide * distance;
-        if (d0 < minSide || d1 < minSide || d2 < minSide)
+        if (legs[0].d < minSide || legs[1].d < minSide || legs[2].d < minSide)
             return { score: 0 };
     }
 
     if (opt.scoring.maxSide !== undefined) {
         const maxSide = opt.scoring.maxSide * distance;
-        if (d0 > maxSide || d1 > maxSide || d2 > maxSide)
+        if (legs[0].d > maxSide || legs[1].d > maxSide || legs[2].d > maxSide)
             return { score: 0 };
     }
 
@@ -203,14 +242,15 @@ export function scoreTriangle(tp, opt) {
     if (!cp)
         return { score: 0 };
 
-    let score = (distance - closingPenalty(cp.d, opt)) * opt.scoring.multiplier;
+    const penalty = closingPenalty(cp.d, opt);
+    let score = opt.scoring.rounding((distance - penalty) * opt.scoring.multiplier);
 
-    return { distance, score, tp, cp };
+    return { distance, score, tp, cp, legs, penalty };
 }
 
 // Upper limit for an out-and-return with 2 TPs (XCLeague) with TPs somewhere in boxes
 export function boundOutAndReturn2(ranges, boxes, opt) {
-    const maxDistance = geom.maxDistance2Rectangles(boxes) * 2;
+    const maxDistance = opt.scoring.rounding(geom.maxDistance2Rectangles(boxes)) * 2;
 
     if (maxDistance < (opt.scoring.minDistance || 0))
         return 0;
@@ -220,16 +260,16 @@ export function boundOutAndReturn2(ranges, boxes, opt) {
         const cp = geom.isTriangleClosed(ranges[0].end, ranges[1].start, maxDistance, opt);
         if (!cp)
             return 0;
-        return (maxDistance - closingPenalty(cp.d, opt)) * opt.scoring.multiplier;
+        return opt.scoring.rounding((maxDistance - closingPenalty(cp.d, opt)) * opt.scoring.multiplier);
     }
 
     // Ranges overlap - bounding is impossible at this stage
-    return maxDistance * opt.scoring.multiplier;
+    return opt.scoring.rounding(maxDistance * opt.scoring.multiplier);
 }
 
 // Score an out-and-return with 2 TPs once the 2 points have been selected
 export function scoreOutAndReturn2(tp, opt) {
-    const distance = tp[0].distanceEarth(tp[1]) * 2;
+    const distance = opt.scoring.rounding(tp[0].distanceEarth(tp[1])) * 2;
 
     if (distance < (opt.scoring.minDistance || 0))
         return { score: 0 };
@@ -238,9 +278,14 @@ export function scoreOutAndReturn2(tp, opt) {
     if (!cp)
         return { score: 0 };
 
-    let score = (distance - closingPenalty(cp.d, opt)) * opt.scoring.multiplier;
+    const penalty = closingPenalty(cp.d, opt);
+    const score = opt.scoring.rounding((distance - penalty) * opt.scoring.multiplier);
+    const legs = [
+        {name: 'TP1 : TP2', start: tp[0], finish: tp[1], d: distance},
+        {name: 'TP2 : TP1', start: tp[1], finish: tp[0], d: distance}
+    ];
 
-    return { distance, score, tp, cp };
+    return { distance, score, tp, cp, legs, penalty };
 }
 
 // Upper limit for an out-and-return with 1 TP (FAI) with a TP somewhere in boxes
@@ -252,7 +297,7 @@ export function boundOutAndReturn1(ranges, boxes, opt) {
         Math.max(boxes[0].x2, boxes[2].x2),
         Math.max(boxes[0].y2, boxes[2].y2),
     );
-    const maxDistance = geom.maxDistance2Rectangles([boxes[1], box2]);
+    const maxDistance = opt.scoring.rounding(geom.maxDistance2Rectangles([boxes[1], box2]));
 
     if (maxDistance < (opt.scoring.minDistance || 0))
         return 0;
@@ -273,32 +318,35 @@ export function boundOutAndReturn1(ranges, boxes, opt) {
             (boxes[0].x2 + boxes[2].x2) / 2,
             (boxes[0].y2 + boxes[2].y2) / 2
         );
-        const realDistance = geom.maxDistance2Rectangles([boxes[1], box2]);
-        return (realDistance - closingPenalty(cp.d, opt)) * 2 * opt.scoring.multiplier;
+        const realDistance = opt.scoring.rounding(geom.maxDistance2Rectangles([boxes[1], box2]));
+        return opt.scoring.rounding((realDistance - closingPenalty(cp.d, opt)) * 2 * opt.scoring.multiplier);
     }
 
     // Ranges overlap - bounding is impossible at this stage
-    return maxDistance * 2 * opt.scoring.multiplier;
+    return opt.scoring.rounding(maxDistance * 2 * opt.scoring.multiplier);
 }
 
 // Score an out-and-return with 1 TPs once the point has been selected
 export function scoreOutAndReturn1(tp, opt) {
-    const distance = Math.max(tp[0].distanceEarth(tp[1]), tp[1].distanceEarth(tp[2]));
+    // Create the second turn point on the middle of the closing line
+    const tp2 = new Point((tp[0].x + tp[2].x) / 2, (tp[0].y + tp[2].y) / 2);
+    const distance = opt.scoring.rounding(tp[1].distanceEarth(tp2));
 
     if (distance < (opt.scoring.minDistance || 0))
         return { score: 0 };
 
-    const d = tp[0].distanceEarth(tp[2]);
-    if (d > opt.scoring.closingDistance(distance, opt))
+    const closing = opt.scoring.rounding(tp[0].distanceEarth(tp[2]));
+    if (closing > opt.scoring.closingDistance(distance, opt))
         return { score: 0 };
 
-    // Create the second turn point on the middle of the closing line
-    const tp2 = new Point((tp[0].x + tp[2].x) / 2, (tp[0].y + tp[2].y) / 2);
-    const realDistance = tp[1].distanceEarth(tp2);
+    const penalty = closingPenalty(closing, opt);
+    const score = (distance - penalty) * 2 * opt.scoring.multiplier;
+    const legs = [
+        {name: 'TP1 : TP2', start: tp[0], finish: tp2, d: distance},
+        {name: 'TP2 : TP1', start: tp2, finish: tp[0], d: distance}
+    ];
 
-    let score = (realDistance - closingPenalty(d, opt)) * 2 * opt.scoring.multiplier;
-
-    return { distance: realDistance, score, tp: [tp[1], tp2], cp: { d, in: tp[0], out: tp[2] } };
+    return { distance, score, tp: [tp[1], tp2], cp: { closing, in: tp[0], out: tp[2] }, legs };
 }
 
 // These implement the FAI Sporting Code, Section 7D, Paragraph 5.2.5
@@ -366,30 +414,35 @@ export function adjustFAICylinders(score, opt) {
     case 'tri':
     case 'fai':
         {
-            const d0 = score.tp[0].distanceEarth(score.tp[1]);
-            const d1 = score.tp[1].distanceEarth(score.tp[2]);
-            const d2 = score.tp[2].distanceEarth(score.tp[0]);
-            score.distance = d0 + d1 + d2 - opt.scoring.cylinders * 2 * 3;
+            score.distance = 0;
+            for (let i = 0; i < score.legs.length; i++) {
+                score.legs[i].d = opt.scoring.rounding(score.tp[i].distanceEarth(score.tp[(i + 1) % score.tp.length]));
+                score.distance += score.legs[i].d;
+            }
+            score.distance -= opt.scoring.cylinders * 2 * 3;
             score.score = score.distance >= (opt.scoring.minDistance || 0) ? 
-                (score.distance - closingPenalty(score.cp.d, opt)) * opt.scoring.multiplier : 0;
+                opt.scoring.rounding((score.distance - closingPenalty(score.cp.d, opt)) * opt.scoring.multiplier) : 0;
         }
         break;
     case 'oar':
         {
-            const realDistance = score.tp[0].distanceEarth(score.tp[1]) - opt.scoring.cylinders;
-            score.score = score.distance >= (opt.scoring.minDistance || 0) ?
-                (realDistance - closingPenalty(score.cp.d, opt)) * 2 * opt.scoring.multiplier : 0;
+            const distance = opt.scoring.rounding(score.tp[0].distanceEarth(score.tp[1])) - opt.scoring.cylinders;
+            score.legs[0].d = score.legs[1].d = distance;
+            score.score = score.minDistance >= (opt.scoring.minDistance || 0) ?
+                opt.scoring.rounding((distance - closingPenalty(distance, opt)) * 2 * opt.scoring.multiplier) : 0;
         }
         break;
     case 'od':
         {
-            score.distance = 0;
             const all = [score.ep.start, score.tp[0], score.tp[1], score.tp[2], score.ep.finish];
-            for (let i = 0; i < all.length - 1; i++)
-                score.distance += all[i].distanceEarth(all[i + 1]);
+            score.distance = 0;
+            for (let i = 0; i < all.length - 1; i++) {
+                score.legs[i].d = opt.scoring.rounding(all[i].distanceEarth(all[i + 1]));
+                score.distance += score.legs[i].d;
+            }
             score.distance -= opt.scoring.cylinders * 2 * 3 + opt.scoring.cylinders * 2;
             score.score = score.distance >= (opt.scoring.minDistance || 0) ?
-                score.distance * opt.scoring.multiplier : 0;
+                opt.scoring.rounding(score.distance * opt.scoring.multiplier) : 0;
         }
         break;
     }